Add awesome-open-source
The awesome-open-source list collects resources how to do open source. This includes fundamentals such as information about open source licensing, but also practical advice on topics such as community building and managing open source in an organization. It's not a list of specific open source projects or applications, but focuses on the general practices and culture open source is based on.